REFACTOR `DependenciesService` to `Sfneal\Dependencies\Utils` namespace
